requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
ExifRead==3.0.0
Pillow==9.4.0
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.7.1
sqlparse==0.4.3
